RG-2542 Reinvent Table Component: move legacy tables
