+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@jetbrains/ring-ui",
-  "version": "7.0.109",
+  "version": "8.0.0-beta.0",
   "description": "JetBrains UI library",
   "author": {
     "name": "JetBrains"
