+++ src/data-list/selection.ts
@@ -1,4 +1,8 @@
-import TableSelection, {type CloneWithConfig, type SelectionItem, type TableSelectionConfig} from '../table/selection';
+import TableSelection, {
+  type CloneWithConfig,
+  type SelectionItem,
+  type TableSelectionConfig,
+} from '../legacy-table/selection';
 
 interface DataListSelectionConfig<T extends SelectionItem> extends TableSelectionConfig<T> {
   partialSelected?: Set<T> | undefined;
