+++ src/legacy-table/row-with-focus-sensor.tsx
