+++ src/legacy-table/selection-adapter.ts
