+++ src/legacy-table/selection.ts
