+++ src/legacy-table/simple-table.tsx
