+++ src/legacy-table/smart-table.tsx
